Make player x and y integers
As uint8_t, the top out of bounds would never be negative.
So if the game was flipped, the player would die too easily.
Now we can check if the player has fallen upward because y can be less than 0.

+++ src/player.cpp
@@ -1,6 +1,6 @@
 #include "player.h"
 
-Player::Player(uint8_t x, uint8_t y, int speedX, int speedY, uint8_t size,
+Player::Player(int x, int y, int speedX, int speedY, uint8_t size,
                float jumpDuration, const unsigned char *image) {
   this->x = x;
   this->y = y;
